+++ x/staking/simulation/operations.go
@@ -9,6 +9,7 @@ import (
 	simappparams "github.com/cosmos/cosmos-sdk/simapp/params"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
 	"github.com/cosmos/cosmos-sdk/x/staking/keeper"
 	"github.com/cosmos/cosmos-sdk/x/staking/types"
@@ -118,24 +119,25 @@ func SimulateMsgCreateValidator(ak types.AccountKeeper, bk types.BankKeeper, k k
 
 		denom := k.GetParams(ctx).BondDenom
 
-		balance := bk.GetBalance(ctx, simAccount.Address, denom).Amount
-		if !balance.IsPositive() {
+		delegatable := bk.SpendableCoins(banktypes.WithVestingLockedBypass(ctx), simAccount.Address)
+		delegatableAmt := delegatable.AmountOf(denom)
+		if !delegatableAmt.IsPositive() {
 			return simtypes.NoOpMsg(types.ModuleName, types.TypeMsgCreateValidator, "balance is negative"), nil, nil
 		}
 
-		amount, err := simtypes.RandPositiveInt(r, balance)
+		selfDelegationAmt, err := simtypes.RandPositiveInt(r, delegatableAmt)
 		if err != nil {
 			return simtypes.NoOpMsg(types.ModuleName, types.TypeMsgCreateValidator, "unable to generate positive amount"), nil, err
 		}
 
-		selfDelegation := sdk.NewCoin(denom, amount)
+		selfDelegation := sdk.NewCoin(denom, selfDelegationAmt)
 
-		account := ak.GetAccount(ctx, simAccount.Address)
-		spendable := bk.SpendableCoins(ctx, account.GetAddress())
-
+		spendable := bk.SpendableCoins(ctx, simAccount.Address)
+		coins, hasNeg, err := calculateAmountAvailableForFees(delegatable, spendable, selfDelegation)
+		if err != nil {
+			return simtypes.NoOpMsg(types.ModuleName, types.TypeMsgCreateValidator, "unable to calculate spendable"), nil, err
+		}
 		var fees sdk.Coins
-
-		coins, hasNeg := spendable.SafeSub(selfDelegation)
 		if !hasNeg {
 			fees, err = simtypes.RandomFees(r, ctx, coins)
 			if err != nil {
@@ -261,24 +263,25 @@ func SimulateMsgDelegate(ak types.AccountKeeper, bk types.BankKeeper, k keeper.K
 			return simtypes.NoOpMsg(types.ModuleName, types.TypeMsgDelegate, "validator's invalid echange rate"), nil, nil
 		}
 
-		amount := bk.GetBalance(ctx, simAccount.Address, denom).Amount
-		if !amount.IsPositive() {
+		delegatable := bk.SpendableCoins(banktypes.WithVestingLockedBypass(ctx), simAccount.Address)
+		delegatableAmt := delegatable.AmountOf(denom)
+		if !delegatableAmt.IsPositive() {
 			return simtypes.NoOpMsg(types.ModuleName, types.TypeMsgDelegate, "balance is negative"), nil, nil
 		}
 
-		amount, err := simtypes.RandPositiveInt(r, amount)
+		toDelegateAmt, err := simtypes.RandPositiveInt(r, delegatableAmt)
 		if err != nil {
 			return simtypes.NoOpMsg(types.ModuleName, types.TypeMsgDelegate, "unable to generate positive amount"), nil, err
 		}
 
-		bondAmt := sdk.NewCoin(denom, amount)
+		toDelegate := sdk.NewCoin(denom, toDelegateAmt)
 
-		account := ak.GetAccount(ctx, simAccount.Address)
-		spendable := bk.SpendableCoins(ctx, account.GetAddress())
-
+		spendable := bk.SpendableCoins(ctx, simAccount.Address)
+		coins, hasNeg, err := calculateAmountAvailableForFees(delegatable, spendable, toDelegate)
+		if err != nil {
+			return simtypes.NoOpMsg(types.ModuleName, types.TypeMsgDelegate, "unable to calculate spendable"), nil, err
+		}
 		var fees sdk.Coins
-
-		coins, hasNeg := spendable.SafeSub(bondAmt)
 		if !hasNeg {
 			fees, err = simtypes.RandomFees(r, ctx, coins)
 			if err != nil {
@@ -286,7 +289,7 @@ func SimulateMsgDelegate(ak types.AccountKeeper, bk types.BankKeeper, k keeper.K
 			}
 		}
 
-		msg := types.NewMsgDelegate(simAccount.Address, val.GetOperator(), bondAmt)
+		msg := types.NewMsgDelegate(simAccount.Address, val.GetOperator(), toDelegate)
 
 		txCtx := simulation.OperationInput{
 			R:             r,
@@ -565,5 +568,29 @@ func SimulateMsgBeginRedelegate(ak types.AccountKeeper, bk types.BankKeeper, k k
 		}
 
 		return simulation.GenAndDeliverTxWithRandFees(txCtx)
+	}
+}
+
+// calculateAmountAvailableForFees compares the delegatable and spendable amounts, and factors in the amount being delegated
+// in order to return the amount of coins that will be available for fees after toDelegate is removed from the account.
+// Returns the coins available for fees, whether the calculation resulting in a negative amount, and any error encountered.
+func calculateAmountAvailableForFees(delegatable, spendable sdk.Coins, toDelegate sdk.Coin) (sdk.Coins, bool, error) {
+	delegatableAmt := delegatable.AmountOf(toDelegate.Denom)
+	spendableAmt := spendable.AmountOf(toDelegate.Denom)
+	switch {
+	case delegatableAmt.Equal(spendableAmt):
+		coins, hasNeg := spendable.SafeSub(toDelegate)
+		return coins, hasNeg, nil
+	case delegatableAmt.GT(spendableAmt):
+		vestingAmt := delegatableAmt.Sub(spendableAmt)
+		if toDelegate.Amount.GT(vestingAmt) {
+			coins, hasNeg := delegatable.SafeSub(toDelegate)
+			return coins, hasNeg, nil
+		} else {
+			return spendable, false, nil
+		}
+	default:
+		// Shouldn't be possible.
+		return nil, false, fmt.Errorf("spendable %q is greater than delegatable %q", spendable, delegatable)
 	}
 }