+++ CHANGELOG.md
@@ -48,9 +48,15 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 * [#581](https://github.com/provenance-io/cosmos-sdk/pull/581) For `MsgMultiSend` and `InputOutputCoins`, allow many inputs when there's a single output.
 
+### Bug Fixes
+
+* [#582](https://github.com/provenance-io/cosmos-sdk/pull/582) Prevent locked coins from being delegated. Coins locked in a vesting account can still be delegated though.
+* [#582](https://github.com/provenance-io/cosmos-sdk/pull/582) Staking simulations can no longer try to delegate coins locked outside of vesting.
+
 ### API Breaking
 
 * [#581](https://github.com/provenance-io/cosmos-sdk/pull/581) The `InputOutputCoins` once again takes in multiple inputs. It returns an error if there are multiple inputs and multiple outputs.
+* [#582](https://github.com/provenance-io/cosmos-sdk/pull/582) The `SimulateFromSeed` function now also returns the last block time simulated.
 
 ---